pyproject.toml
[tool.poetry]
name = "portfolio"
version = "0.1.0"
description = "automation of the integration of jsonresumes & latex templates"
authors = ["Santiago Armstrong <santiagoarmstrong@gmail.com>"]

[tool.poetry.dependencies]
python = ">=3.8,<3.11"
pydantic = "^1.8.2"
python-dotenv = "^0.19.1"
GitPython = "^3.1.27"



[tool.poetry.dev-dependencies]
pytest = "^5.2"
pylint = "^2.11.1"
black = "^21.9b0"
bandit = "^1.7.1"
mypy = "^0.930"
isort = "^5.10.1"
flake8 = "^4.0.1"
ipykernel = "^6.15.1"

[tool.poetry.scripts]
prezzi = "prezzi.main:main"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
